refactor (user) : 유즈케이스 테스트용이하게 수정
