--- src/main/java/com/backend/immilog/user/application/usecase/CompanyInquireUseCase.java
@@ -1,7 +1,0 @@
-package com.backend.immilog.user.application.usecase;
-
-import com.backend.immilog.user.application.result.CompanyResult;
-
-public interface CompanyInquireUseCase {
-    CompanyResult getCompany(Long userSeq);
-}
