--- src/main/java/com/backend/immilog/user/application/usecase/CompanyRegisterUseCase.java
@@ -1,10 +1,0 @@
-package com.backend.immilog.user.application.usecase;
-
-import com.backend.immilog.user.application.command.CompanyRegisterCommand;
-
-public interface CompanyRegisterUseCase {
-    void registerOrEditCompany(
-            Long userSeq,
-            CompanyRegisterCommand command
-    );
-}
